Move context Getter Setter
