+++ api/handler/room/wait_room.go
@@ -6,7 +6,6 @@ import (
 	"net/http"
 
 	"github.com/go-playground/validator/v10"
-	"github.com/pollenjp/gameserver-go/api/auth"
 	"github.com/pollenjp/gameserver-go/api/entity"
 	"github.com/pollenjp/gameserver-go/api/handler"
 	"github.com/pollenjp/gameserver-go/api/service"
@@ -47,7 +46,7 @@ func (wr *WaitRoom) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	userId, ok := auth.GetUserId(ctx)
+	userId, ok := service.GetUserId(ctx)
 	if !ok {
 		handler.RespondJson(ctx, w, &handler.ErrResponse{
 			Message: "failed to get user id from context",
