+++ api/handler/user/update_user.go
@@ -6,9 +6,9 @@ import (
 	"net/http"
 
 	"github.com/go-playground/validator/v10"
-	"github.com/pollenjp/gameserver-go/api/auth"
 	"github.com/pollenjp/gameserver-go/api/entity"
 	"github.com/pollenjp/gameserver-go/api/handler"
+	"github.com/pollenjp/gameserver-go/api/service"
 )
 
 // TODO: convert to //go:generate when writing tests
@@ -27,7 +27,7 @@ type UpdateUser struct {
 
 func (ru *UpdateUser) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
-	userId, isOk := auth.GetUserId(ctx)
+	userId, isOk := service.GetUserId(ctx)
 	if !isOk {
 		handler.RespondJson(ctx, w, &handler.ErrResponse{
 			Message: "failed to get user id from token",
